Converted frame widget handlers to use typescript
