--- src/main/javascript/modules/widget/frame/bootstrap.js
@@ -1,16 +1,0 @@
-/**
- * Module to initialise the game frame script bindings.
- */
-module.exports = function (scriptManager) {
-	var modules = [
-		require('./experience-counter'),
-		require('./ribbon'),
-		require('./minimap'),
-		require('./options'),
-		require('./confirm')
-	];
-
-	for (var i in modules) {
-		modules[i].init(scriptManager);
-	}
-};
