+++ src/main/javascript/modules/widget/frame/index.ts
@@ -1,0 +1,5 @@
+import './confirm';
+import './experience-counter';
+import './minimap';
+import './options';
+import './ribbon';
